+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 1.1 (TBD)
 * Pressing C to adjust pivot distance now ignores backfaces when backface culling is enabled in viewport settings.
+* Added camera roll mouse action. Can press R to reset roll.
 
 # 1.0 (2019-10-01)
 * Initial release